# Go/APG1_Golang_Boot_camp/APG1_Golang_Boot_camp/Go_Day06/src/ex01/docs/swagger.yaml
basePath: /
host: localhost:8888
info:
  contact: {}
  description: This is a simple blog API with admin panel, markdown support, and pagination.
  title: My Blog API
  version: "1.0"
paths:
  /api/articles:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Post a new article from the admin panel
      parameters:
      - description: Article Title
        in: formData
        name: title
        required: true
        type: string
      - description: Article Content
        in: formData
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: nil
      summary: Post a new article
      tags:
      - articles
  /api/articles/{id}:
    get:
      consumes:
      - application/json
      description: Get a single post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Post HTML
          schema:
            type: json
        "404":
          description: Not Found
          schema:
            type: json
      summary: Show a single post
      tags:
      - articles
  /api/auth/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authenticate user and generate a JWT token.
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /api/auth/logout:
    post:
      description: Clear the authentication token.
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Logout
      tags:
      - auth
swagger: "2.0"